added the bfgs source file, which can be compiled as a test program if you pass in -DEXECUTE to gcc.
the maxbfgs source wraps up the appropriate
functions from libEmu, and is a part of libEmu
itself so that you can do maximisation with fn
pointers and such using the bfgs routine laidout in
bfgs.c.

This is nice but theres some slightly tricky nested
functioning and pointering so you should be pretty
sure what you're doing. The evalfn and gradfn rely
on inheriting the arguments from maxWithBFGS which
compiles but isnt actually tested yet. (fingers
crossed)