Changed ExternalModel to get information about the model from the executable itself
The ExternalModel was reading a configuration file that held the
number of parameters, their names and range, the number of outputs and
their names, and the path to the executable.

Instead of using a configuration file, ExternalModel now gets the
model configuration information from the executable itself. This
requires the executable launched by ExternalModel to write information
about the model to stdout.